Solve UI bugs on Employees and add Details
